+++ src/main/java/com/javaweb/project/repository/PostRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Set;
 
-public interface PostRepository extends JpaRepository<Post, Long>, PostCustom {
+public interface PostRepository extends JpaRepository<Post, Integer>, PostCustom {
     public void deleteById(Long id);
     public boolean existsBySlugIgnoreCase(String slug);
     public List<Post> findByTitleIsContainingIgnoreCase(String title);
