+++ src/main/java/com/javaweb/project/repository/TagRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TagRepository extends JpaRepository<Tag, Long> {
+public interface TagRepository extends JpaRepository<Tag, Integer> {
     public boolean existsBySlug(String slug);
     public Tag findBySlug(String slug);
 }
