+++ src/main/java/com/javaweb/project/service/CommentService.java
@@ -3,6 +3,6 @@ package com.javaweb.project.service;
 import com.javaweb.project.dto.request.CommentRequestDTO;
 
 public interface CommentService {
-    public void addCommentToPost(Long postId, CommentRequestDTO commentRequestDTO);
-    public void deleteComment(Long id);
+    public void addCommentToPost(Integer postId, CommentRequestDTO commentRequestDTO);
+    public void deleteComment(Integer id);
 }
