+++ src/main/java/com/javaweb/project/service/PostService.java
@@ -14,11 +14,11 @@ import java.util.Set;
 public interface PostService {
     public List<PostDTO> findAllBlogs();
 //    public List<PostDTO> findBlogsByTitleOrAuthorName(String title, String authorName);
-    public void updateBlogPost(Long id, UpdatePostRequest request);
-    public void deleteBlogPostById(Long id);
-    public PostDTO findBlogById(Long id);
+    public void updateBlogPost(Integer id, UpdatePostRequest request);
+    public void deleteBlogPostById(Integer id);
+    public PostDTO findBlogById(Integer id);
 
-    public PostDetailDTO getAPostDetail(Long id);
+    public PostDetailDTO getAPostDetail(Integer id);
     public void createNewBlogPost(CreatePostRequest request);
     public List<PostDTO> getAllMyBlog();
     public List<PostDTO> findPostsByTitle(String title);
